Switch explicitly to 32 bit types wherever possible in cuda
This is a contribution from Kevin Jaget and team 900.  Kevin says this
is faster, and makes the code work better on nvcc and older versions of
cuda.  I'm not surprised it helps.  We also very much blow up if any
size is bigger than 4 GB already, this doesn't make any of those worse.

Change-Id: I752209f66fc337c2d42ce03dc5fb6fd89eb63e8b
Signed-off-by: Kevin Jaget <kjaget@gmail.com>
Signed-off-by: Austin Schuh <austin.linux@gmail.com>