Refactor ImageCallback to expose CameraImage messages
Now that we have images of different types, we don't actually want any
of the conversion logic to run before the GPU apriltag code.  We really
just want the raw data pointer and the type.

Split the conversion code into ImageCallback, and move all the sync
logic to CameraImageCallback.  Since we now know about image types, make
ImageCallback a bit smarter and explode when asked to produce an image
type it doesn't know how to.

Change-Id: Ic0b5727951792edbbfd88f99c0a9761d4eb9b9d2
Signed-off-by: Austin Schuh <austin.linux@gmail.com>