Skip creating Senders for Non-Replayed Channels
When using ReplayChannels with LogReader, Senders are still being
created even if messages will never be sent on them. This change skips
that logic.

This is needed since we run into issues when replaying on target where
we're creating Senders for channels that either don't exist or changed
in the configuration on the target.

Change-Id: I1a931f1968c38051c682e372f641787ab415a835
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>