Move Result<> Run() methods to new overloads
I238e7aa6b82dc3d7938630b2b0ab96eb30b573eb made the Run*() methods of the
SimulatedEventLoopFactory return Result<>'s to support us propagating
up errors from both applications and from AOS internals. In actually
implementing some of these errors, I realized that by simply swapping
out the existing methods that would make it so that any existing code
using these methods would have to wrap every single call to Run() with a
CheckExpected() in order to get the same behavior that they had before
(given that the errors that I am implementing were previously
 CHECKs/LOG(FATAL)s).

This would likely have affected on the order of 1000 call-sites in the
code that I am aware of. As such, this change reverts the Run*() calls
to their original semantics (aborting on failure) and creates new
overloads that return Result<>'s and which specify [[nodiscard]] to
ensure that the Result<>'s are used.

The main alternatives would have been to provide some sort of run-time
means of specifying the preferred behavior (via command line flag or
argument to the Run() call or some sort of configuration parameter on
the SimulatedEventLoopFactory). However, my experience with these sorts
of behaviors is that they are error-prone as we could still easily end
up with call-sites that are not properly handling the return values and
accidentally turn on non-fatal failures or the such.

Change-Id: I399f7429a159d00efa3c972c7d8375eca8097fe6
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>