aos: Delete some unused shm-related code
I can't find any users of this code so I figured it's simplest to
delete it. I stumbled across it when I was trying to understand our
`mmap` setup for AOS fetchers. I am looking into making the underlying
memory for fetchers non-pageable. That lets us avoid a copy for CUDA.

Change-Id: Ic7d85c2cd76a9507713b846a5355909fe946a0f4
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>