Improve error message on failure to allocate static flatbuffer
When we had a static flatbuffer that took up more space than was
available in the AOS channel, we tended to produce error messages that
did not actually tell you which channel, how much we were trying to
allocate, etc.

This fixes the error message for the case where the initial allocation
of a static flatbuffer fails. It does not directly do anything for cases
where the user may attempt to grow the buffer beyond the allowable
bounds (in these cases the static flatbuffer API generally returns an
error rather than crashing).

This also adds tests for this behavior for both the raw and static
flatbuffer senders.

Change-Id: I1814e0c8ed233ac55d555de1898f2450fe410464
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>