Support regexs `hostnames` option for `Node`s
This makes it so that instead of specifying a literal string or set of
strings as your hostname you may instead specify a set of regex's.

Note that this does technically alter the semantics of existing
configurations; however, I am not aware of any configurations that are
likely to be affected.

Change-Id: I336c44029b6f97ca9700ab21a4eb226d2a5781f0
Signed-off-by: James Kuszmaul <jabukuszmaul+collab@gmail.com>