aos: Add new helper to strip schemas from a Configuration
I am looking to make use of the `configuration` channel in Foxglove.
Currently it's _huge_ and causes noticeable lag. Especially when you
render it in the Raw Message panel. This patch will let me strip the
schemas from that channel so that Foxglove is less laggy. There's
still a bit of lag, but it's much better now.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I632eab46082562427e5c4a90a1e8402b661aab0f