Turn on --bfbs-comments for flatbuffer compilation (#745)
This makes it so that any comment lines starting with a `///`
in .fbs files will be included in the reflection schema.

This will be useful for situations where we want to be able to pass
through documentation from the `.fbs` files to the generated schemas
for use in more processed documentation.

In particular, for handling PII fields I would like to both have an attribute
on each field that is PII, but then also take any comments for that field/table
and attach them to generated documentation which describes which fields
in our messages are PII.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I7f7413459980812fc79143dc8aeee16e00d0e2c4