aos: Add some conversion metadata for MCAP files
I have some code that needs to perform a channel lookup in the
configuration. Since I'm not using canonical channel names most of the
time, the naive lookup will fail. For example, depending on the node
you're replaying from, the "/aos aos.starter.Status" channel actually
maps to a different channel.

This patch adds a new metadata message into the MCAP log under the
topic name `log_conversion_metadata`. For now, messages on this topic
will just contain the node on which the log was replayed.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I1410b83b1dbda445c8f78015435741846c61798c