Annotate malloc'ing sections of sctp_lib
There are some sections of sctp_lib.cc that run in the realtime
codepaths currently but which do malloc. Ideally, they would not malloc
at all, but in the interim we will mark just these as non-realtime so
that the malloc hooks do get run during the rest of execution.

Tested on dblite; message_bridge_* no longer crash.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I398bd3a0213b7668eec9ec17045c606e51d90761