Use "newest" schema in AOS config merging
In general our config merging logic is supposed to take the "last" state
of any given *thing* as the state to provide in the resulting
configuration. However, for the reflection schemas we were instead
taking the *first* schema which we encountered. This was causing issues
for certain log-replay scenarios where we would attempt to insert an
updated reflection schema into the AOS config but not end up with it
actually taking effect.

I do not believe the "only use the first schema" was even the original
intent of this code, because it was effectively relying on
`absl::btree_set::emplace()` to immediately destroy the newer schemas
when you attempted to insert them.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I0591778fe22d23a566f3f4e4b23e7e4462579f2a