Fix flatbuffers::Verify() method for <4 byte buffers
Apparently flatbuffers::Verify() never checked that the input buffer you
provided was actually large enough to validly read the root offset from
the start of the buffer (to the best of my knowledge it does correctly
handle buffers larger than 4 bytes).

This was caught by the address sanitizers in the nightly.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I18a4e964fd72b71191995857eadea65f54e2fa8a