Add a CHECK in static flatbuffers to prevent aos::realtime crash.
The static flatbuffers allow specifying a static_length in the
flatbuffer (fbs) files.

Previously, if the FromFlatbuffer method was used and caused
more elements to be allocated than was defined in the
static_length attribute, we would get a short stack track with
no indication what caused aos::realtime to crash.

This change adds a check so we can see a useful stack trace.

Change-Id: Ibcc6d1d5f433c5758941005d75230ebac6bcbbd8
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>