Potentially fix CI retry behavior on instance termination
The docs say this should be working, but it isn't.  The only difference
I can see between what the docs show and what we are doing is that the
docs specify it for each step, and we specify it once.  Let's try
switching over and see if that fixes it.

Change-Id: I52867b0b182b90e6d123f17d25d3312769474fa9
Signed-off-by: Austin Schuh <austin.linux@gmail.com>