Add support to aos_dump/aos_send for raw bfbs
It is quite handy to be able to send the exact contents of a bfbs file
out as a message, and to save it.  For this particular use case, I
wanted to save an image to a file, and repeatedly send it back out for
testing.  aos_dump can now save the message, and aos_send can send it.

This also makes unit testing with the raw contents of a message easier,
since you can just dump the contents with aos_dump to a file, and load
it back in the test.

Change-Id: I9fd8c15ccb7f0a2606748574e3328ca242cc91f8
Signed-off-by: Austin Schuh <austin.linux@gmail.com>