Support standard-compliant JSON generation in FlatbufferToJson
Currently, there are several scenarios in which the FlatbufferToJson
code will generate non-compliant JSON. In particular:
* In the presence of non-finite floating point numbers, `nan`, and `inf`
  will be generated (without quotes).
* In the presence of non-UTF-8 strings, `\xXX` escape sequences will be
  generated in strings.

This adds a `use_standard_json` option which causes `FlatbufferToJson`
to generate valid JSON in both of those cases (in particular,
`nan`/`inf` will be wrapped in quotation marks, with the assumption
that consumers know to interpret the fields as numbers, and non-UTF-8
strings will be rendered as if they were byte arrays).

This requires adding support to the JSON parsing code to ensure that we
acn round-trip any JSON generated this way.

This also makes minor incidental clean-ups to the reflection-based
FlatbufferToJson code (flatbuffer_introspection.cc), because it was not
actually generating strings entirely correctly (in particular, it did
not generate proper escape sequences for new-line characters and the
such when rendering vectors of strings).

The most complex logic change required in this is that, in order to
support parsing vectors of strings where the strings may be byte arrays,
I had to remove some of the assumptions that we would only ever
be nested one vector deep. Fortunately, this does not seem to have been
too substantial a change.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I9770e239657f155de82991d1f0f74f7b164aa637