Output "-0.0" instead of "-0" in FlatbufferToJson
Most JSON parsers---including our own---interpret a "-0" as an integer
when parsing and so lose the sign bit when parsing the JSON, which
results in information loss. Special-case -0.0 in particular when
serializing flatbuffers to JSON.

It may be appropriate to consider forcing all float/double JSON
serialization to include a decimal point to universally remove
confusion, but I believe that -0.0 is the only source of actual
ambiguity.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I34cc0804f47453d239d8eb6f86f1c3b8b88c8ae4