aos: Allow jinja2_template to use command line flags
Occasionally I find myself wanting to iterate on experiments more
quickly than the "edit, save, build" loop. Instead, I want to tweak a
parameter or two directly on the `bazel build` command line. That
turns out to be a decent amount faster.

One component in enabling this is to put the command line flags into
source code or config files etc. The `jinja2_template` rule is the
fastest way to accomplish this. With that I can quickly tweak settings
in an AOS configuration JSON file without having to touch it at all.

This came up during my experiments to validate that enabling SCTP
priority has the desired effect.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I419c19588d1de93a4835ebc46bfcc0cee9a94468