aos: Allow log_stats to ignore the beginning of a log
I have some logs where the network went down a few times at the
beginning of the log. That caused some latency spikes that I want to
ignore in the statistics. This patch adds the ability to skip X
seconds at the beginning of the log by specifying `--skip_seconds=X`
on the command line.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Ic3498372efc118f22afc30f78d7ca75e693f5187