aos: Set a minimum SO_SNDBUF for the message bridge
On our internal setup, we see timestamps getting dropped on the
message bridge client.

    2025-03-11_20-42-12.042731636 (62522.464833694sec) /0/aos aos.message_bridge.ClientStatistics: {
      "connections": [
        {
          "node": {
            "name": "vpu0-1a"
          },
          "state": "CONNECTED",
          "received_packets": 44820306,
          "monotonic_offset": -1976682271,
          "partial_deliveries": 0,
          "boot_uuid": "35e7f731-7322-4d32-b60f-d4acaa5bc96d",
          "connected_since_time": 41367725088,
          "connection_count": 2,
          "timestamp_send_failures": 623730
        },
        {
          "node": {
            "name": "vpu0-4a"
          },
          "state": "CONNECTED",
          "received_packets": 44816685,
          "monotonic_offset": 4208820036,
          "partial_deliveries": 0,
          "boot_uuid": "ac352e5e-6227-4da3-82c8-66cc897a741b",
          "connected_since_time": 46836165504,
          "connection_count": 2,
          "timestamp_send_failures": 623981
        }
      ]
    }

Those `timestamp_send_failures` values keep incrementing with every
statistics message. Turning on some verbose logging, we see this:

    $ /tmp/message_bridge_client --disable_ipv6 --wants_sctp_authentication -vmodule=sctp_lib=1 --nodie_on_malloc
    ...
    W0311 21:02:28.918219  395954 sctp_lib.cc:419] sendmsg on sctp socket failed: Resource temporarily unavailable [11]
    W0311 21:02:28.918382  395954 sctp_lib.cc:419] sendmsg on sctp socket failed: Resource temporarily unavailable [11]
    W0311 21:02:29.018364  395954 sctp_lib.cc:419] sendmsg on sctp socket failed: Resource temporarily unavailable [11]
    W0311 21:02:29.018472  395954 sctp_lib.cc:419] sendmsg on sctp socket failed: Resource temporarily unavailable [11]
    W0311 21:02:29.118497  395954 sctp_lib.cc:419] sendmsg on sctp socket failed: Resource temporarily unavailable [11]
    W0311 21:02:29.118594  395954 sctp_lib.cc:419] sendmsg on sctp socket failed: Resource temporarily unavailable [11]
    ...

As far as I can tell, the issue here is that the message buffer is
sized to fit one message of the biggest size. For timestamp messages,
pretty much all timestamp messages are roughly the same size. That
means that sending a single timestamp message will fill up the buffer.

The current buffer is sized to ~1448 bytes:

    $ /tmp/message_bridge_client --disable_ipv6 --wants_sctp_authentication
    ...
    I0311 21:02:34.474510  396248 message_bridge_client_lib.cc:171] Max write message size for all servers is 1448
    ...

One option here is to change the retry logic to not wait 100ms to
retry _all_ timestamp messages. Those too will fail after a single
timestamp message is sent out.

Another option is to adjust the send buffer size. That's what this
patch here does. That lets more than a single message fit into the
kernel send buffer. Based on my experiments, it completely eliminated
the send failures for timestamps.

In order to keep the logic for `--force_wmem_max` working, I moved the
logic from `message_bridge_server_lib.cc` into `sctp_lib.cc`. That
makes it easier to manage and adds the functionality to the client
code as well.

This patch likely has no effect on the server that uses this code.
Messages are likely to already be 64KiB or larger. The server side
buffer also gets sized linearly with the number of connections a
channel has.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I31b4fa0865d6f7d3636641b3006c32ebc43f0d68