aos: Allow the foxglove client to send messages
I am looking to make my life easier to controlling a system from
foxglove. This patch enables the server side of this functionality.

One slightly awkward thing (but probably good from a "security" point
of view) is that the server needs to pre-specify which topics the
client is allowed to publish on. That means we need to create AOS
senders on startup.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Ic7d8fdc233a60309625ed08259cd80104a0e96a6