aos: Let live foxglove connections find the node they're connected to
In order to let the clients do some channel mapping, they have to know
what node they're connected to. Otherwise the `Configuration` is
difficult to parse.

This patch adds the necessary information so that the clients can
perform the lookup that C++ applications automatically do with
`AddWatcher` and friends.

There's a potential argument to combine the `LogConversionMetadata`
with this new `LiveMetadata` message, but I couldn't quite convince
myself of this. So for now they're still separate.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: Ic3d3b092a4c30b55f4e08b60d5b431f605566997