aos: Allow LOG statements to use the simulated clock
When debugging issues in simulation (e.g. via log replay), it's often
useful to see the simulated timestamps in LOG statements. The wall
clock doesn't _really_ have a lot of information during simulation.

This patch lets the user specify `--simulated_log_clocks` so that
the simulated time gets printed as well as the current node name.

This came up when trying to reproduce a segfault by rerunning against
some logs.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I855063111aba11be1d931a9f04ef42906af923e9