aos: Only print one stack trace per `LOG(FATAL)`
I'm fixing the stack unwinding in signal handlers. But that means that
`LOG(FATAL)` statements now print 2 stack traces. One for the LOG
statement itself and another when `std::abort` is called (due to
SIGABRT).

This patch fixes the issue by not printing the stack trace during the
SIGABRT handler if a `LOG(FATAL)` stack trace has already been
printed. The hooks for this are already present in abseil. It's not
obvious to me why it hasn't been enabled upstream.

Since I'm adding a new patch to abseil, I took this opportunity to
regenerate the existing patch into a proper patch.

I added a few tests for the stack traces. Two of them are disabled,
however, until I get around to actually making stack traces work for
signal handlers.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Change-Id: I68c1606456c2d9883e3db62d42e3ad7bcc526419