aos: Detect more use-after-free bugs with flatbufffers
When using the `FlatbufferFixedAllocatorArray`, there's currently no
guard against re-using pointers after calling `.Reset()` on the
object.

For example, this is currently fine:

    FlatbufferFixedAllocatorArray<T, kSize> array;
    // Build the message.
    ...
    const T* ptr1 = &array.message();

    array.Reset();

    // Build the message again. Possibly with very different contents.
    ...
    const T* ptr2 = &array.message();

    // Using ptr1 can invoke undefined behaviour sometimes.

Unfortunately, it seems like there are no hooks into the sanitizers
for custom allocators. Especially ones that work on arena-backed
memory like `FlatbufferFixedAllocatorArray`.

This patch works around the limitation by using an `std::vector` so
that the underlying data is re-allocated every time.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Co-authored-by: Philipp Schrader <philipp.schrader@bluerivertech.com>
Change-Id: Ibae8c9788f54b8e7b7eb0c94d8aedda4b114deea