(aos): Provide convenient AnnotateBinary method
I keep finding myself wanting to be able to inspect the underlying
memory of a flatbuffer and having to constantly copy the relevant
annotation code around. This modifies things to make it so that you can
conveniently use the flatbuffer binary annotator (including making it so
that it does not need to write/read files from the local
filesystem).

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Co-authored-by: James Kuszmaul <91565539+jameskuszmaul-brt@users.noreply.github.com>
Change-Id: I8fbc83fbc6ec59d7ea4ff92eacd9b39b2ec5d80f