aos: Break out `log_to_mcap`'s channel dropping logic
I want to add an expensive interceptor on the `LogReader` object via
`AddBeforeSendCallback()`. But that logic will live in a custom
version of `log_to_mcap`. When the channels are dropped, however,
there's no  point in having this expensive interceptor. It's
detrimental to the log conversion runtime. This means I only want to
register the interceptor if the channel will _not_ be dropped.

This patch splits out the channel dropping logic so that I can easily
re-use it in my custom `log_to_mcap` binary.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Co-authored-by: Philipp Schrader <philipp.schrader@bluerivertech.com>
Change-Id: Iacc77e3b86ec562147da139ac2da97ca8126ec08