aos: Make the `Result` a bit more ergonomic
The goal is to make the `aos::Result` type a little more ergonomic.
The current `Status` alias doesn't play well with the underlying
`std::expected` API. For example:

    Status status = PerformAction();
    if (status.has_value()) {
       // There's a status, right?
    }

The `has_value()` member method here is will return true if
`PerformAction()` was successful. It will return false if
`PerformAction()` returned an error. But a "status" is semantically
trying to convey an "error" state. So using `has_value()` is very
confusing in this instance.

This patch here aims to make `Result` more ergonomic:
* Making the value type default to `void`. This lets users write
  `Result<>` instead of `Result<void>`.
* Adding a few helper functions to make call sites more readable.

An example of a helper function:

    Result<> result = PerformAction();
    if (HasError(result)) {
        // There's an error:
    }
    if (IsOk(result)) {
        // There's no error:
    }

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Co-authored-by: Philipp Schrader <philipp.schrader@bluerivertech.com>
Change-Id: I712152989172286a12dc308970dde451f799677f