aos: Fix undefined behavior in static flatbuffer `SetString`
With the current code, the new test encounters this issue:

    [ RUN      ] StaticFlatbuffersTest.EmptyString
    external/aos/aos/flatbuffers/static_vector.h:812:32: runtime error: null pointer passed as argument 2, which is declared to never be null
    external/amd64_sysroot_jetpack5/usr/include/string.h:44:28: note: nonnull attribute specified here

The `memcpy` call should not be called with a nullptr as an argument,
even if the size is zero.

The code now only calls `memcpy` if the size is non-zero.

This came up while writing a test case for the `SetStringOrDie`
helper.

Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>
Co-authored-by: Philipp Schrader <philipp.schrader@bluerivertech.com>
Change-Id: Ibf90a9a2d56e8af3771f47585440c10f9f5a387b